\name{as.lm.calibintstat}
\alias{as.lm.calibintstat}
\title{ Fitting Linear Models}
\description{ attempts to coerce its argument into a linear model. }
\usage{\method{as.lm}{calibintstat}(object,...)}
\arguments{
\item{...}{}
\item{object}{}
}
\value{
\item{ object of class "lm".}{}
}
\author{Witold Wolski \email{wolski@molgen.mpg.de}}
\keyword{misc}
