\name{cal}
\docType{data}                     
\alias{cal}
\title{Calibration masses}
\usage{data(cal)} 
\description{A massvector with calibration masses. (mainly trypsin autoproteolysis peaks)}                                
\author{Witold E. Wolski}         
\keyword{datasets}     
               