\name{calibspline}
\alias{calibspline}
\title{ Constructor}
\description{ Returns object of class calibspline. It is used by the function \code{getextcalib.massvector}. }
\usage{calibspline(ispl,mv,error,theo)}
\arguments{
\item{error}{ vector with errors.}
\item{ispl}{ object of class "smooth.spline"}
\item{mv}{ massvector}
\item{theo}{ vector with theoretical masses}
}
\author{Witold Wolski \email{wolski@molgen.mpg.de}}
\seealso{\code{\link{getextcalib.massvector}},
}
\keyword{misc}
