\name{getcoorBruker}
\alias{getcoorBruker}
\title{ Coordinates on Sample support.}
\description{ Gets the coordinates on sample support from the default infor parameter (e.g. \code{0_A1_1SRef}). }
\usage{getcoorBruker(infop)}
\arguments{
\item{infop}{ path to folder.}
}
\author{Witold Wolski \email{wolski@molgen.mpg.de}}
\keyword{misc}
