\name{histLengths}
\alias{histLengths}
\title{ Histograms}
\description{ Histogram of the massvector lengths. }
\usage{histLengths(x,main=info(x),xlab="length of massvectors",...)}
\arguments{
\item{...}{ further parameters.}
\item{main}{ info(x)}
\item{x}{ massvectorlist}
\item{xlab}{ length of massvectors.}
}
\author{Witold Wolski \email{wolski@molgen.mpg.de}}
\seealso{\code{\link[base]{hist}},
}
\keyword{misc}
