\name{image.massvector}
\alias{image.massvector}
\title{ Display a Color Image}
\description{ Creates a grid of colored or gray-scale rectangles with colors  corresponding to the mass differences within the peaklist  or within two peaklists. }
\usage{\method{image}{massvector}(x,mv2,error=NULL,ppm=FALSE,col=topo.colors(100),...)}
\arguments{
\item{...}{}
\item{col}{ a list of colors such as that generated by `rainbow',`heat.colors', `topo.colors', `terrain.colors' or similar functions.}
\item{error}{ up to which mass difference display the differences.}
\item{mv2}{ massvector}
\item{ppm}{}
\item{x}{ massvector}
}
\author{Witold Wolski \email{wolski@molgen.mpg.de}}
\seealso{\code{\link{plot.massvector}},
\code{\link{hist.massvector}},
}
\examples{
 data(mv2)
 data(mv1)
 image(mv1,mv2)
 image(mv1,mv2,error=500)
}
\keyword{misc}
