\name{image.mlist}
\alias{image.mlist}
\title{ Display a Color Image}
\description{ Creates a grid of colored or gray-scale rectangles with colors  corresponding to the values in 'z'.  This can be used to display  three-dimensional or spatial data aka "images". This is a generic  function. }
\usage{\method{image}{mlist}(x,what="",col=terrain.colors(100),...)}
\arguments{
\item{...}{}
\item{col}{ a list of colors such as that generated by 'rainbow', 'heat.colors', 'topo.colors', 'terrain.colors' or similar functions.}
\item{what}{ what value to display on the image.}
\item{x}{ object of class mlist. (e.g}
}
\author{Witold Wolski \email{wolski@molgen.mpg.de}}
\examples{
 data(mvl)
 image(mvl,what="lengthmv")
}
\keyword{misc}
