\name{image.mylistobj}
\alias{image.mylistobj}
\title{ image}
\description{ image }
\usage{\method{image}{mylistobj}(x,...)}
\arguments{
\item{...}{}
\item{x}{ object of class mylistobj}
}
\details{ mylistobj provides basic functionality for objects implemented using list to store the attributes. }
\author{Witold Wolski \email{wolski@molgen.mpg.de}}
\keyword{misc}
