\name{length.massvector}
\alias{length.massvector}
\title{ Length of Massvector}
\description{ get the Length of a Massvector }
\usage{\method{length}{massvector}(x,...)}
\arguments{
\item{...}{}
\item{x}{ massvector}
}
\author{Witold Wolski \email{wolski@molgen.mpg.de}}
\examples{
 data(mv1)
 length(mv1)
}
\keyword{misc}
