\name{max.massvector}
\alias{max.massvector}
\title{ Maxima and Minima}
\description{ Returns the maxima and minima of the masses and intensities }
\usage{\method{max}{massvector}(mv,...)}
\arguments{
\item{...}{}
\item{mv}{ massvector}
}
\value{
\item{ named array with maxima of the columns of mv.}{}
}
\author{Witold Wolski \email{wolski@molgen.mpg.de}}
\seealso{\code{\link{min.massvector}},
}
\keyword{misc}
