\name{min.massvector}
\alias{min.massvector}
\title{ Maxima and Minima}
\description{ Returns the maxima and minima of the masses and intensities }
\usage{\method{min}{massvector}(mv,...)}
\arguments{
\item{...}{}
\item{mv}{ massvector}
}
\value{
\item{ named array with minima of the columns of mv.}{}
}
\author{Witold Wolski \email{wolski@molgen.mpg.de}}
\seealso{\code{\link{max.massvector}},
}
\keyword{misc}
