\name{mv1}
\docType{data}                                               
\alias{mv1}
\alias{mv2}                                      
\title{Massvector with experimentally determined peptide masses.}           
\usage{data(mv1)}   
\description{}                                                                                     
\author{Witold E. Wolski}                                   
\keyword{datasets}                                              
