\name{plot.mylistobj}
\alias{plot.mylistobj}
\title{ plot}
\description{ plot }
\usage{\method{plot}{mylistobj}(x,...)}
\arguments{
\item{...}{}
\item{x}{ object of class mylistobj}
}
\details{ mylistobj provides basic functionality for objects implemented using list to store the attributes. }
\author{Witold Wolski \email{wolski@molgen.mpg.de}}
\keyword{misc}
