\name{print.massvector}
\alias{print.massvector}
\title{ Print massvector}
\description{ `print' prints its argument and returns it invisibly (via `invisible(x)') }
\usage{\method{print}{massvector}(x,quote=FALSE,...)}
\arguments{
\item{...}{}
\item{quote}{}
\item{x}{ massvector}
}
\author{Witold Wolski \email{wolski@molgen.mpg.de}}
\seealso{\code{\link[base]{print}},
}
\examples{
 data(mv1)
 print(mv1)
}
\keyword{misc}
