\name{print.massvectorlist}
\alias{print.massvectorlist}
\title{ Print massvectorlist}
\description{ `print' prints its argument and returns it invisibly (via `invisible(x)') }
\usage{\method{print}{massvectorlist}(x,...)}
\arguments{
\item{...}{}
\item{x}{ massvectorlist}
}
\author{Witold Wolski \email{wolski@molgen.mpg.de}}
\seealso{\code{\link[base]{print}},
}
\examples{
 data(mvl)
 print(mvl)
}
\keyword{misc}
