\name{readBruker.massvector}
\alias{readBruker.massvector}
\title{ Read peaklist.xml}
\description{ Read Bruker Daltonic peaklist.xml file format. }
\usage{\method{readBruker}{massvector}(object,filep,infop,...)}
\arguments{
\item{...}{}
\item{filep}{ path and filename of file to load.}
\item{infop}{ info field. If missing the info field is extracted out of the file path.}
\item{object}{ object of class massvector.}
}
\author{Witold Wolski \email{wolski@molgen.mpg.de}}
\keyword{misc}
