\name{summary.caliblist}
\alias{summary.caliblist}
\title{ Massvector Summaries}
\description{ Generates a summary: min , max etc for the data.frame generated by as.data.frame.massvectorlist }
\usage{\method{summary}{caliblist}(object,...)}
\arguments{
\item{...}{}
\item{object}{ caliblist}
}
\author{Witold Wolski \email{wolski@molgen.mpg.de}}
\seealso{\code{\link{as.data.frame.caliblist}},
\code{\link[base]{summary}},
}
\examples{
 data(mvl)
 mvl<-mvl[1:100]
 data(cal)
 test <- getintcalib(mvl,cal,error=500)
 summary(test)
 summary(test)
}
\keyword{misc}
