\name{summary.massvector}
\alias{summary.massvector}
\title{ Massvector Summaries}
\description{ Generates a summary: min, max etc. }
\usage{\method{summary}{massvector}(object,...)}
\arguments{
\item{...}{}
\item{object}{ massvector}
}
\author{Witold Wolski \email{wolski@molgen.mpg.de}}
\seealso{\code{\link{summary}},
}
\keyword{misc}
