\name{summary.massvectorlist}
\alias{summary.massvectorlist}
\title{ Massvector Summaries}
\description{ Generates a summary: min, max etc for the data.frame generated by as.data.frame.massvectorlist }
\usage{\method{summary}{massvectorlist}(object,...)}
\arguments{
\item{...}{ further arguments}
\item{object}{ massvectorlist}
}
\author{Witold Wolski \email{wolski@molgen.mpg.de}}
\seealso{\code{\link{summary.massvector}},
\code{\link{as.data.frame.massvectorlist}},
}
\examples{
 data(mvl)
 summary(mvl)
}
\keyword{misc}
