\name{summary.mlist}
\alias{summary.mlist}
\title{ mist Summaries}
\description{ Generates a summary for the data.frame generated by as.data.frame.mlist }
\usage{\method{summary}{mlist}(object,...)}
\arguments{
\item{...}{ further arguments}
\item{object}{ mlist}
}
\author{Witold Wolski \email{wolski@molgen.mpg.de}}
\seealso{\code{\link{summary.massvector}},
\code{\link{as.data.frame.mlist}},
}
\examples{
 data(mvl)
 summary(mvl)
 data(mvl)
 mvl<-mvl[1:100] 
 data(cal)
 test <- getintcalib(mvl,cal,error=500)
 summary(test)
}
\keyword{misc}
