\name{writeF.massvectorlist}
\alias{writeF.massvectorlist}
\title{ Write massvectorlist}
\description{ Write massvectorlist to File }
\usage{\method{writeF}{massvectorlist}(object,path,file=experiment(object),ext="txt",...)}
\arguments{
\item{...}{}
\item{ext}{ file extension. default txt.}
\item{file}{ file name. default experiment(object)}
\item{object}{ massvectorlist.}
\item{path}{ path to directory.}
}
\author{Witold Wolski \email{wolski@molgen.mpg.de}}
\seealso{\code{\link{readF.massvectorlist}},
\code{\link{readF.massvector}},
\code{\link{readBruker.massvector}},
\code{\link{readBruker.massvectorlist}},
}
\keyword{misc}
